requirements.txt
joblib
flask
scikit-learn
requests

Cython>=0.22
cmdstanpy==0.9.5
pystan>=2.14
numpy>=1.15.4
pandas>=1.0.4
matplotlib>=2.0.0
LunarCalendar>=0.0.9
convertdate>=2.1.2
holidays>=0.10.2
setuptools-git>=1.2
python-dateutil>=2.8.0
tqdm>=4.36.1
